# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Store Platform API
  version: 1.0.0
  description: API REST da plataforma de vendas pós-pagas
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      summary: Listar usuários internos
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de usuários
  /admin/users/invitations:
    post:
      tags: [admin]
      summary: Convidar usuário interno
      security: [{ bearerAuth: [] }]
      responses:
        '204':
          description: Convite criado
  /admin/roles:
    get:
      tags: [admin]
      summary: Papéis internos
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Papéis
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: store_session